Fix various bugs
Bugs in Pause Menu:

Fix:
When the cursor is over the bow and arrow, the arrow should be
displayed, not the bow.

Fix:
Every time the pause menu is open, the cursor starts in the top-left
position.

Fix:
Link’s green-square indicator should not appear in room 1-1 (the
basement) or any debug rooms.

Fix:
Resetting does not reset the room covers in the pause menu’s map.
The game cannot be reset in any game state other than Playing and
JumpMap.